Move ShuffleShardSeed to util package.
